major refactoring; added learning rate adjustment that needs fixing

+++ src/distances.jl
@@ -17,7 +17,7 @@ pbcdistance(p1, p2, box)
 Computes 3D periodic boundary distance between two points
 """
 function pbcdistance(p1, p2, box)
-    R2::Float64 = 0.
+    R2::Float32 = 0.
     @fastmath @inbounds for i in 1:3
         R2 += pbcdx(p1[i], p2[i], box[i])^2
     end
@@ -32,7 +32,7 @@ Builds the distance matrix for a given
 configuration
 """
 function builddistanceMatrix(conf, box)
-    distanceMatrix = zeros(Float64, length(conf), length(conf))
+    distanceMatrix = zeros(Float32, length(conf), length(conf))
     @inbounds for i in 1:length(conf)
         @inbounds for j in 1:length(conf)
             distanceMatrix[i,j] = pbcdistance(conf[i], conf[j], box)
